diabetes_api/requirements.txt
amqp==5.1.1
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==3.2
uWSGI==2.0.19.1
django-nameko==0.8.2
djangorestframework==3.13.1
dnspython==1.16.0
eventlet==0.33.3
greenlet==2.0.2
idna==3.4
kombu==5.3.1
MarkupSafe==2.1.3
mock==5.1.0
nameko==2.14.1
packaging==23.1
path==16.7.1
path.py==12.5.0
psycopg2-binary==2.9.6
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
Werkzeug==2.3.7
wrapt==1.15.0
